requirements.txt
openai
retry
python-dotenv
discord
requests